delete debug project + code clean
