+++ MvtWatermark/MvtWatermark.sln
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NoDistortionWatermarkMetrics", "NoDistortionWatermarkMetrics\NoDistortionWatermarkMetrics.csproj", "{D09E99B8-7B27-4A89-B393-7BA390F47EB5}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ParameterValues", "ParameterValues\ParameterValues.csproj", "{645B8188-FE49-4257-A6BF-2FFD20073FAE}"
 	ProjectSection(ProjectDependencies) = postProject
 		{20DD51F1-4547-4C98-A47A-45B4DE955E27} = {20DD51F1-4547-4C98-A47A-45B4DE955E27}
@@ -22,8 +20,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Distortion", "Distortion\Distortion.csproj", "{8E415897-2A85-49C9-A9CE-34ADC6CE7CCD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DistortionTry", "DistortionTry\DistortionTry.csproj", "{E12EF0C3-78F4-497B-8E01-6E8CE27EE296}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DebugProj", "DebugProj\DebugProj.csproj", "{CA466783-EC01-4852-A76B-C7B43C6FAD15}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,10 +35,6 @@ Global
 		{E168E5B1-52B5-41E8-87D7-71C1AC5667C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E168E5B1-52B5-41E8-87D7-71C1AC5667C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E168E5B1-52B5-41E8-87D7-71C1AC5667C2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D09E99B8-7B27-4A89-B393-7BA390F47EB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D09E99B8-7B27-4A89-B393-7BA390F47EB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D09E99B8-7B27-4A89-B393-7BA390F47EB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D09E99B8-7B27-4A89-B393-7BA390F47EB5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{645B8188-FE49-4257-A6BF-2FFD20073FAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{645B8188-FE49-4257-A6BF-2FFD20073FAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{645B8188-FE49-4257-A6BF-2FFD20073FAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -55,10 +47,6 @@ Global
 		{E12EF0C3-78F4-497B-8E01-6E8CE27EE296}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E12EF0C3-78F4-497B-8E01-6E8CE27EE296}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E12EF0C3-78F4-497B-8E01-6E8CE27EE296}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CA466783-EC01-4852-A76B-C7B43C6FAD15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CA466783-EC01-4852-A76B-C7B43C6FAD15}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CA466783-EC01-4852-A76B-C7B43C6FAD15}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CA466783-EC01-4852-A76B-C7B43C6FAD15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
