+++ MvtWatermark/MvtWatermark/NoDistortionWatermark/Documentation.txt
@@ -1,4 +1,8 @@
-﻿	При встраивании ЦВЗ, если одна из подпоследовательностей будет состоять целиком из False, то есть при переводе в десятичную систему счисления 
+﻿	
+	Здесь будет описание алгоритма и его особенности, которые потом перемещу в Readme уже на английском, а пока так
+
+
+	При встраивании ЦВЗ, если одна из подпоследовательностей будет состоять целиком из False, то есть при переводе в десятичную систему счисления 
 получится 0, NonStaticMapboxTileWriterWm сгенерирует исключение.
 
 	При извлечении ЦВЗ, в изначально созданном BitArray (размера: [количество тайлов в дереве] * [options.Nb]), состоящем исключительно из False,
